refactor: move *Route.tsx to src/app/routes/ (closes #128)
Routes are composition-layer code — they wire features together and live
in app/, not inside a single feature folder. Moving them eliminates the
invisible ESLint carve-out that allowed *Route.tsx to cross-feature-import;
the no-cross-feature-imports ban is now invariant with no exceptions.

- Move 9 *Route.tsx + 6 *Route.test.tsx from src/features/<feature>/ to
  src/app/routes/. Former relative intra-feature imports become @/features/.
- Update lazy imports in src/app/routes.tsx and direct imports in
  src/app/routes.test.tsx.
- Drop the ignores: ['src/features/**/*Route.tsx'] carve-out and the
  companion block from eslint.config.js.
- Update README to document app/routes/ as where routes live.
