+++ src/app/routes/KidSequenceRoute.tsx
@@ -1,13 +1,12 @@
 import { type JSX, useEffect } from 'react';
 import { Navigate, useNavigate, useParams } from 'react-router-dom';
 
+import { KidSequence } from '@/features/kid-sequence/KidSequence';
 import { clearLastBoard, setLastBoard } from '@/lib/lastBoard';
 import { isNotFoundError, useBoard } from '@/lib/queries/boards';
 import { KidModeGate } from '@/ui/KidModeGate/KidModeGate';
 
-import { KidChoice } from './KidChoice';
-
-export const KidChoiceRoute = (): JSX.Element | null => {
+export const KidSequenceRoute = (): JSX.Element | null => {
   const { boardId = '' } = useParams();
   const { data: board, error } = useBoard(boardId);
   const navigate = useNavigate();
@@ -20,7 +19,7 @@ export const KidChoiceRoute = (): JSX.Element | null => {
   if (!board) return null;
   return (
     <KidModeGate onExitConfirmed={() => navigate(`/boards/${board.id}/edit`)}>
-      {(requestExit) => <KidChoice board={board} onExit={requestExit} />}
+      {(requestExit) => <KidSequence board={board} onExit={requestExit} />}
     </KidModeGate>
   );
 };
