+++ src/app/routes/ParentHomeRoute.tsx
@@ -1,6 +1,8 @@
 import { type JSX, useEffect, useState } from 'react';
 import { Navigate, useNavigate } from 'react-router-dom';
 
+import { NewBoardModal } from '@/features/parent-home/NewBoardModal';
+import { ParentHome } from '@/features/parent-home/ParentHome';
 import { useKidModeNav } from '@/layouts/useKidModeNav';
 import { useParentNav } from '@/layouts/useParentNav';
 import { getLastBoard, hasAutoLaunched, kidPathFor, markAutoLaunched } from '@/lib/lastBoard';
@@ -8,9 +10,6 @@ import { useBoards, useCreateBoard } from '@/lib/queries/boards';
 import { useKids } from '@/lib/queries/kids';
 import { accentForIndex } from '@/theme/tokens';
 import { NewKidModal } from '@/ui/NewKidModal/NewKidModal';
-
-import { NewBoardModal } from './NewBoardModal';
-import { ParentHome } from './ParentHome';
 
 export const ParentHomeRoute = (): JSX.Element => {
   const navigate = useNavigate();
