+++ src/app/routes/SettingsRoute.tsx
@@ -1,11 +1,10 @@
 import type { JSX } from 'react';
 
+import { AccountSection } from '@/features/settings/AccountSection';
+import { DeleteAccountSection } from '@/features/settings/DeleteAccountSection';
 import { ParentShell } from '@/layouts/ParentShell';
 import { useKidModeNav } from '@/layouts/useKidModeNav';
 import { useParentNav } from '@/layouts/useParentNav';
-
-import { AccountSection } from './AccountSection';
-import { DeleteAccountSection } from './DeleteAccountSection';
 
 export const SettingsRoute = (): JSX.Element => {
   const onNav = useParentNav();
